refactor(app): extract dashboard tile editor modal
